+++ src/constants/urls.ts
@@ -1,0 +1,6 @@
+export const basePaths = {
+  users: {
+    development: 'http://www.mocky.io/v2/',
+    production: 'http://www.mocky.io/v2/',
+  },
+};
