--- src/utils/constants.ts
@@ -1,6 +1,0 @@
-export const basePaths = {
-    users: {
-      development: 'http://www.mocky.io/v2/',
-      production: 'http://www.mocky.io/v2/',
-    },
-  };
