+++ src/utils/networkLayer/httpHeaders.ts
@@ -1,0 +1,9 @@
+export const basicHeaders = {
+  'Content-Type': 'application/json',
+};
+
+export const basicSecureHeaders = {
+  'Content-Type': 'application/json',
+  'Authorization': '',
+  'Organization': ''
+};
